schema_mapping/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.9.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.54.0
google-generativeai==0.8.1
requests==2.31.0
asyncpg==0.29.0
pymongo==4.6.1

